refactor: extract swap alloyed asset to tokens to mod
